+++ src/components/collection/add-collection-project-pop.js
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React, { useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import { partition, uniqBy } from 'lodash';
 
@@ -6,7 +6,7 @@ import { getAllPages } from 'Shared/api';
 import Loader from 'Components/loader';
 import { PopoverWithButton, PopoverDialog, PopoverSection, PopoverInfo, InfoDescription } from 'Components/popover';
 import TextInput from 'Components/inputs/text-input';
-import ResultsList from 'Components/containers/results-list';
+import ResultsList, { ScrollResult, useActiveIndex } from 'Components/containers/results-list';
 import Emoji from 'Components/images/emoji';
 import ProjectResultItem from 'Components/project/project-result-item';
 import { useTrackedFunc } from 'State/segment-analytics';
@@ -14,6 +14,7 @@ import { createAPIHook } from 'State/api';
 import { useCurrentUser } from 'State/current-user';
 import { useAlgoliaSearch } from 'State/search';
 
+import useDebouncedValue from '../../hooks/use-debounced-value';
 import { useNotifications, AddProjectToCollectionMsg } from '../../presenters/notifications';
 
 function parseQuery(query) {
@@ -39,43 +40,12 @@ const useTeamProjects = createAPIHook(async (api, teamId) => {
   return null;
 });
 
-function useActiveIndex(items, onSelect) {
-  const [activeIndex, setActiveIndex] = useState(-1);
-  useEffect(() => {
-    setActiveIndex(-1);
-  }, [items]);
-  const onKeyDown = (e) => {
-    if (e.key === 'ArrowUp') {
-      e.preventDefault();
-      setActiveIndex((prev) => {
-        if (prev < 0) {
-          return items.length - 1;
-        }
-        return prev - 1;
-      });
-    } else if (e.key === 'ArrowDown') {
-      e.preventDefault();
-      setActiveIndex((prev) => {
-        if (prev === items.length - 1) {
-          return -1;
-        }
-        return prev + 1;
-      });
-    } else if (e.key === 'Enter') {
-      e.preventDefault();
-      if (items[activeIndex]) {
-        onSelect(items[activeIndex]);
-      }
-    }
-  };
-  return { activeIndex, onKeyDown };
-}
-
 function AddCollectionProjectPop({ collection, togglePopover, addProjectToCollection }) {
   const [query, setQuery] = useState('');
   const parsedQuery = parseQuery(query);
+  const debouncedQuery = useDebouncedValue(query, 200);
   const { topResults, project: retrievedProjects, status } = useAlgoliaSearch(
-    parsedQuery,
+    debouncedQuery,
     {
       notSafeForKids: false,
       filterTypes: ['project'],
@@ -144,7 +114,11 @@ function AddCollectionProjectPop({ collection, togglePopover, addProjectToCollec
       {visibleProjects.length > 0 && (
         <PopoverSection>
           <ResultsList scroll items={visibleProjects}>
-            {(project, i) => <ProjectResultItem project={project} active={i === activeIndex} onClick={() => onClick(project)} />}
+            {(project, i) => (
+              <ScrollResult active={i === activeIndex}>
+                <ProjectResultItem project={project} active={i === activeIndex} onClick={() => onClick(project)} />
+              </ScrollResult>
+            )}
           </ResultsList>
         </PopoverSection>
       )}
